Pass IoT hub information from ARM template to setup script (#238)
Refactor setup.sh to manage all the input values in an easier way, i.e. parameters are passed by name, not by position. More parameters can be passed in now, which can be useful in future development. In particular pass in parameters needed to create the link to IoT Hub metrics in the Azure portal.